requirements.txt
aiobotocore==2.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
asttokens==3.0.1
attrs==25.4.0
boto3==1.41.5
botocore==1.41.5
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cloudpickle==3.1.1
comm==0.2.3
cryptography==46.0.5
decorator==5.2.1
executing==2.2.1
filelock==3.25.0
frozenlist==1.8.0
fsspec==2025.12.0
idna==3.11
importlib_resources==6.5.2
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
jmespath==1.1.0
joblib==1.5.3
jupyterlab_widgets==3.0.16
llvmlite==0.46.0
matplotlib-inline==0.2.1
modin==0.37.1
multidict==6.7.1
numba==0.64.0
numpy==2.4.2
pandas==2.3.1
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
retrying==1.4.2
s3fs==2025.12.0
s3transfer==0.15.0
scikit-learn==1.7.2
scipy==1.17.1
setuptools==80.10.2
shap==0.50.0
six==1.17.0
slicer==0.0.8
snowflake-connector-python==4.3.0
snowflake-ml-python==1.29.0
snowflake-snowpark-python==1.45.0
sortedcontainers==2.4.0
sqlparse==0.5.5
stack-data==0.6.3
threadpoolctl==3.6.0
tomlkit==0.14.0
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
wcwidth==0.6.0
wheel==0.46.3
widgetsnbextension==4.0.15
wrapt==1.17.3
xgboost==3.2.0
yarl==1.23.0
